change parameter type to double

+++ src/utils/Matrix.java
@@ -1,7 +1,7 @@
 package utils;
 
 public class Matrix {
-    public static void printMatrix(int[][] A) {
+    public static void printMatrix(double[][] A) {
         for (int i = 0; i < A.length; i++) {
             for (int j = 0; j < A[i].length; j++) {
                 System.out.print(A[i][j] + " ");
